legend for globe and hover component update
