+++ app/landingglobe/page.tsx
@@ -3,6 +3,7 @@
 import React, { useState } from "react";
 import dynamic from "next/dynamic";
 import MainPageNav from "./popups/LandingGlobeNav";
+import GlobeLegend from "./popups/GlobeLegend";
 
 const Globe = dynamic(() => import("./R3fGlobeScene"), {
   ssr: false,
@@ -21,19 +22,7 @@ export default function LandingGlobe() {
     <div className="relative w-full h-screen">
       <Globe />
       <MainPageNav open={isOpen} setIsOpen={setIsOpen} />
-      <div className="absolute border-border border bg-background top-4 right-2 hidden md:block w-34 h-24 p-2 rounded-lg">
-        Hover a country and see the investment needed in
-        <ul className="list-none">
-          <li>
-            <span className="inline-block w-3 h-3 bg-[hsl(45,100%,50%)] mx-1 rounded-sm" />{" "}
-            Solar
-          </li>
-          <li>
-            <span className="inline-block w-3 h-3 bg-[hsl(195,85%,65%)] mx-1 rounded-sm" />{" "}
-            Wind
-          </li>
-        </ul>
-      </div>
+      <GlobeLegend />
     </div>
   );
 }
